requirements.txt
pandas>=1.3.0
numpy>=1.21.0
statsmodels>=0.13.0
optuna>=3.0.0
scikit-learn>=1.0.0
nba_api>=1.1.0
matplotlib>=3.4.0
seaborn>=0.11.0
haversine>=2.5.0
pyarrow>=12.0.0
PyYAML>=6.0
mplcursors>=0.7